Updates for clover.

+++ benchmarks/blur/blur_kernels.cl
@@ -1,15 +1,15 @@
 
 kernel
 void
-blur_hor(global uchar* im1, global uchar* im0, global double* bvec, 
+blur_hor(global uchar* im1, global uchar* im0, global float* bvec, 
         int ww, int hh, int sigma)
 /* @spec blur_hor(ww, hh, sigma) global_size(0, 1) */
 {
     int ii = get_global_id(0);
     int jj = get_global_id(1);
    
-    int    rr = 3 * sigma;
-    double pp = 0.0;
+    int   rr = 3 * sigma;
+    float pp = 0.0f;
 
     for (int kk = -rr; kk <= rr; ++kk) {
         int jj0 = clamp(jj + kk, 0, ww - 1);
@@ -21,15 +21,15 @@ blur_hor(global uchar* im1, global uchar* im0, global double* bvec,
 
 kernel
 void
-blur_ver(global uchar* im1, global uchar* im0, global double* bvec, 
+blur_ver(global uchar* im1, global uchar* im0, global float* bvec, 
         int ww, int hh, int sigma)
 /* @spec blur_ver(ww, hh, sigma) global_size(0, 1) */
 {
     int ii = get_global_id(0);
     int jj = get_global_id(1);
    
-    int    rr = 3 * sigma;
-    double pp = 0.0;
+    int   rr = 3 * sigma;
+    float pp = 0.0f;
 
     for (int kk = -rr; kk <= rr; ++kk) {
         int ii0 = clamp(ii + kk, 0, hh - 1);
